Refactor: 티켓 상세 정보 주소 추가

See merge request s08-final/S08P31D108!103

+++ AOS/app/src/main/java/com/team/parking/data/model/ticket/TicketBuyConfirmResponse.kt
@@ -8,6 +8,6 @@ data class TicketBuyConfirmResponse(
     val parkingRegion: String,
     val ptHas: Int,
     val ptLose: Int,
-    val shaId: Int,
+    val shaId: Long,
     val type: Int
 )
