+++ AOS/app/src/main/java/com/team/parking/data/model/ticket/TicketSellConfirmResponse.kt
@@ -5,6 +5,6 @@ data class TicketSellConfirmResponse(
     val parkingRegion: String,
     val ptToEarn: Int,
     val sellConfirm: Boolean,
-    val shaId: Int,
-    val ticketId: Int
+    val shaId: Long,
+    val ticketId: Long
 )
